requirements.txt

psutil
PyYAML
